requirements.txt
fastmcp>=3.0,<4.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
streamlit>=1.32,<2.0
langchain>=1.0,<2.0
langchain-google-genai>=4.0,<5.0
